+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [1.5.3] - 2026-01-01
+
+- Venus: Add support for Venus A (VNSA) and Venus D (VNSD) device types
+- Fix loop when all configured devices have invalid/unknown device types.
 
 ## [1.5.2] - 2025-10-03